Pig Game
My version of the Pig Game from The Complete JavaScript Course 2018: Build Real Projects! on udemy by Jonas Schmedtmann.

Some simple changes like locking the text field that allows you to change the score when the game has ended.  Also added a div that appears when you lose the round and locks the controls for 1.5 seconds before removing the div and allowing the play to continue.